src/Agro-Server/requirements.txt
asgiref==3.10.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
google-api-core==2.28.0
google-auth==2.41.1
google-cloud-bigquery==3.38.0
google-cloud-storage==2.18.2
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
pandas==2.2.2
gunicorn
dotenv
gunicorn